LayoutTests/inspector/uisourcecode-revisions-expected.txt

Tests revision support in UISourceCode.

Bug 97669

Running: testAddRevisionsRevertToOriginal
First revision added.
UISourceCode content: content2
    revisions: 'content2'
Second revision added.
UISourceCode content: content4
    revisions: 'content2', 'content4'
Reverted to original.
UISourceCode content: <script content>
    revisions: 'content2', 'content4', '<script content>'

Running: testAddRevisionsRevertAndClearHistory
First revision added.
UISourceCode content: content2
    revisions: 'content2'
Second revision added.
UISourceCode content: content4
    revisions: 'content2', 'content4'
Reverted and cleared history.
UISourceCode content: <script content>
    revisions: 

Running: testAddRevisionsRevertToPrevious
First revision added.
UISourceCode content: content2
    revisions: 'content2'
Second revision added.
UISourceCode content: content4
    revisions: 'content2', 'content4'
Reverted to previous revision.
UISourceCode content: content2
    revisions: 'content2', 'content4', 'content2'

Running: testRequestContentAddRevisionsRevertToOriginal
Content requested.
UISourceCode content: <script content>
    revisions: 
First revision added.
UISourceCode content: content2
    revisions: 'content2'
Second revision added.
UISourceCode content: content4
    revisions: 'content2', 'content4'
Reverted to original.
UISourceCode content: <script content>
    revisions: 'content2', 'content4', '<script content>'

Running: testRequestContentAddRevisionsRevertAndClearHistory
Content requested.
UISourceCode content: <script content>
    revisions: 
First revision added.
UISourceCode content: content2
    revisions: 'content2'
Second revision added.
UISourceCode content: content4
    revisions: 'content2', 'content4'
Reverted and cleared history.
UISourceCode content: <script content>
    revisions: 

Running: testRequestContentAddRevisionsRevertToPrevious
Content requested.
UISourceCode content: <script content>
    revisions: 
First revision added.
UISourceCode content: content2
    revisions: 'content2'
Second revision added.
UISourceCode content: content4
    revisions: 'content2', 'content4'
Reverted to previous revision.
UISourceCode content: content2
    revisions: 'content2', 'content4', 'content2'
